+++ retrievalmarket/network/network.go
@@ -2,7 +2,7 @@ package network
 
 import (
 	"github.com/filecoin-project/go-fil-components/retrievalmarket"
-	peer "github.com/libp2p/go-libp2p-peer"
+	"github.com/libp2p/go-libp2p-core/peer"
 )
 
 type RetrievalQueryStream interface {
@@ -10,6 +10,7 @@ type RetrievalQueryStream interface {
 	WriteQuery(retrievalmarket.Query) error
 	ReadQueryResponse() (retrievalmarket.QueryResponse, error)
 	WriteQueryResponse(retrievalmarket.QueryResponse) error
+	Close() error
 }
 
 type RetrievalDealStream interface {
@@ -19,6 +20,7 @@ type RetrievalDealStream interface {
 	WriteDealResponse(retrievalmarket.DealResponse) error
 	ReadDealPayment() (retrievalmarket.DealPayment, error)
 	WriteDealPayment(retrievalmarket.DealPayment) error
+	Close() error
 }
 
 type RetrievalReceiver interface {
