+++ retrievalmarket/network/query_stream.go
@@ -1,9 +1,9 @@
 package network
 
 import (
-	cborutil "github.com/filecoin-project/go-cbor-util"
 	"github.com/filecoin-project/go-fil-components/retrievalmarket"
-	"github.com/filecoin-project/go-fil-components/retrievalmarket/impl/impl_types"
+	retrievalimpl "github.com/filecoin-project/go-fil-components/retrievalmarket/impl"
+	"github.com/filecoin-project/go-fil-components/shared/cborutil"
 	"github.com/filecoin-project/go-fil-components/shared/tokenamount"
 	"github.com/ipfs/go-cid"
 	p2pnet "github.com/libp2p/go-libp2p-core/network"
@@ -15,8 +15,6 @@ type queryStream struct {
 	s p2pnet.Stream
 }
 
-var _ RetrievalQueryStream = (*queryStream)(nil)
-
 func (qs queryStream) ReadQuery() (retrievalmarket.Query, error) {
 	panic("implement me")
 }
@@ -27,11 +25,11 @@ func (qs queryStream) WriteQuery(q retrievalmarket.Query) error {
 		return err
 	}
 
-	return cborutil.WriteCborRPC(qs.s, &impl_types.OldQuery{Piece: cid})
+	return cborutil.WriteCborRPC(qs.s, &retrievalimpl.OldQuery{Piece: cid})
 }
 
 func (qs queryStream) ReadQueryResponse() (retrievalmarket.QueryResponse, error) {
-	var oldResp impl_types.OldQueryResponse
+	var oldResp retrievalimpl.OldQueryResponse
 	if err := oldResp.UnmarshalCBOR(qs.s); err != nil {
 		log.Warn(err)
 		return retrievalmarket.QueryResponseUndefined, err
@@ -48,6 +46,8 @@ func (qs queryStream) ReadQueryResponse() (retrievalmarket.QueryResponse, error)
 func (qs queryStream) WriteQueryResponse(retrievalmarket.QueryResponse) error {
 	panic("implement me")
 }
+
+var _ RetrievalQueryStream = (*queryStream)(nil)
 
 func (qs queryStream) Close() error {
 	return qs.s.Close()
